+++ src/index.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import RebelBaseApp from './RebelBaseApp';
-
-
 import './index.css';
 import 'materialize-css/dist/css/materialize.min.css'
 import reportWebVitals from './reportWebVitals';
